feat: Update ready listener to set bot activity
This commit updates the ready listener to set the bot's activity to display the number of hackers in the server. It uses the `setActivity` method from the Discord.js library and sets the activity type to "Watching". The activity is updated every 15 seconds. This change improves the user experience by providing real-time information about the server activity.